Move Navigation components to MainPage folder in UI
