+++ ui/apps/platform/src/Containers/MainPage/Header/Masthead.tsx
