+++ ui/apps/platform/src/Containers/MainPage/Header/UserMenu.js
