--- ui/apps/platform/src/Containers/Navigation/NavigationSideBar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NavigationSideBar';
