+++ ui/apps/platform/src/utils/orchestratorComponents.ts
@@ -1,0 +1,13 @@
+import { RestSearchOption } from 'services/searchOptionsToQuery';
+
+export const orchestratorComponentsOption: RestSearchOption[] = [
+    {
+        value: 'Orchestrator Component:',
+        type: 'categoryOption',
+    },
+    {
+        value: 'false',
+    },
+];
+
+export const ORCHESTRATOR_COMPONENTS_KEY = 'showOrchestratorComponents';
